separate out DeltaSharingHistoryManager, build passed waiting for test
